video-automation/requirements.txt
Pillow>=10.0
moviepy>=2.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
python-dotenv>=1.0
requests>=2.28
